server/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
pdfplumber
python-docx
requests
python-dotenv
